Inserts minority types with batch insertion style
i.e. relationship types that have less relationships than whatever
the batch size for the importer is. This is because there's a certain
amount of overhead executing the stages for a relationship type and
if there are few relationships of a type then it's not worth it and
inserting using batch-insert (i.e. random access style) is more
efficient.

+++ community/kernel/src/main/java/org/neo4j/unsafe/impl/batchimport/CalculateDenseNodesStage.java
@@ -67,7 +67,7 @@ public class CalculateDenseNodesStage extends Stage
     /*
      * @see RelationshipTypeCheckerStep#getRelationshipTypes(int)
      */
-    public Object[] getRelationshipTypes( int belowOrEqualToThreshold )
+    public Object[] getRelationshipTypes( long belowOrEqualToThreshold )
     {
         return typer.getRelationshipTypes( belowOrEqualToThreshold );
     }
